Add client caching for FileSystem factory
There are two new inner classes within the FileSystem interface. One is the cache which represents an object that stores references to FileSystem client instances. The other class is a key object which is used to retrieve filesystem instances from the cache.

Both should be invisible to the user. If a user does not want any caching then they should be using the respective #create method to avoid retreiving a cached object.