Remove the need for fs context from AbstractFileSystem
It's odd to need a reference to both the FileSystem client and context from within the AbstractFileSystem. This change removes the FileSystemContext field from the object and introduces the appropriate methods within the FileSystem client.

The major changes are that FileSystem now extends closeable and that we have a new getBlockLocations function.

We've added implementations for close because the underlying FileSystemContext represents a closeable resource which should be cleaned up appropriately when a FileSystem is removed from the cache. However because the context is a protected field we need to expose the close method to the user.

The getBlockLocations function has been implemented as a user facing API in the Alluxio client because the Hadoop API needs a way to retrieve information about the block location without access to the FileSystemContext. The next best place which would have access to the context is within the client itself.

Associated tests have been added and some modified for the new methods and functionality.

One caveat which we need to be addressed is how we should handle closing resources on the thread-safe client if there are operations currently in progress or that are blocked and waiting on resources. This is still up for discussion.