Make netty EventLoopGroup held in the FileSystemContext
This allows us to respect the users' configuration of the netty worker group. By moving it into the FIleSystemContext we are able to keep the reference to the group and all operations through a single client will use the same reference. With FileSystem caching enabled we will only use a minimal amount of extra resources if a user creates a new FileSystem instance.