remove timeouts from equality and hash code
It's not required to put the health check timeout and shutdown timeout inside of the equality and hashCode functions for the ChannelKey.

The shutdown of a channel will always immediately return if there are other references present. The last key to shut down the channel will respect the timeout however which is what we want. Similarly for health check which is only used when acquiring the channel - it will come from the key and is not stored with the channel so it is not required in those functions.