Remove static reference to EventLoopGroup in DefaultBlockWorkerClient (#8341)
* Make netty EventLoopGroup held in the FileSystemContext

This allows us to respect the users' configuration of the netty worker group. By moving it into the FIleSystemContext we are able to keep the reference to the group and all operations through a single client will use the same reference. With FileSystem caching enabled we will only use a minimal amount of extra resources if a user creates a new FileSystem instance.

* checkstyle fixes

* shut down workerGroup and FIleSystemContext close

* use metrics app id in netty thread name

* fix checkstyle

* rename netty pool thread

Renaming to the new format gives the developer more information about the thread. Useful for debugging jstacks.