Diligently release resources
Previously, none of the web servers closed their FIleSystem client. The
AsyncCacheRequestManager created it's own FileSystemContext upon
instantiation. This has been changed such that the GrpcDataServer
is now reponsible for creating and closing the FileSystemContext that
the AsyncCacheRequestManager uses. Lastly, the ClientPool,
LocalAlluxioClusterResource, and PollingMasterInquireClientTest all had
cases where resources may not be released, so I made sure to close them
properly.