Diligently release resources - web servers, clients, and ports (#8372)
Previously, none of the web servers closed their FIleSystem client. The
AsyncCacheRequestManager created it's own FileSystemContext upon
instantiation. This has been changed such that the GrpcDataServer
is now reponsible for creating and closing the FileSystemContext that
the AsyncCacheRequestManager uses. Lastly, the ClientPool,
LocalAlluxioClusterResource, and PollingMasterInquireClientTest all had
cases where resources may not be released, so I made sure to close them
properly.

Due to the stop() of a cluster now closing all of the clients within the client pool, any test which utilizes a client after stopping or restarting a cluster now needs to get a new client before making any requests.