[#8366] Move FileSystemUtils into FileSystem (#8369) (#8408)
* Move FileSystemUtils into FileSystem

Tests were integrated into FileSystemIntegrationTest.

* Leave primitive methods in FileSystem client

Compound operations have been moved or left inside of FileSystemUtils

* Update perist command to add timeout

During testing RunOperation can't seem to create files any more unless
we default the CreatePOptions recursive value to true. So this has been
set for CreateFile.

Related shell utilities and tests have been updated

* Add persistence options to client API

ScheduleAsyncPersistencePOptions already existed. It needs to be
exposed on the client so that users can pass options to it in the future
if desired

* Remove unused options sWriteBoth

* Fix confusing argument in createUnpersistedFiles

* Address minor typos in PersistCommand output

* move checkConsistency to respective command class

checkConsistency isn't used anywhere except inside the command. It doesn't make sense to leave it inside FileSystemUtils because of it's requirement for a FileSystemContext. It doesn't belong in the client either.

* Rename waitPersisted to persistAndWait

* Test both option forms

Added test for getIntArg as well as PersistCommand

* style issues

* fix javadoc nits

* Improvements to persist exception handling and logging

These changes make the persist mirror more closely how the other client functions handle errors. Some variables were also renamed in order to make the function more similar to others. A debug logging statement was added as well.

* rename variables

- `client` moved to `masterClient` in BaseFileSystem
- `uri` changed to `path` in FileSystem