Fix Alluxio Yarn integration (#8420)
* No need to exclude transport from common module.

There are classes in alluxio-core-common that depend on grpc, grpc is only included in alluxio-core-transport, so it has to be included, otherwise, ClassNotFound issues will happen. Previously, we exclude alluxio-core-protobuf because of some shading hacks which no longer work for alluxio-core-transport.

* Give current work directory a higher priority in classpath.

Alluxio uses a newer version of Guava, an older version of Guava also exists in the YARN_APPLICATION_CLASSPATH, so without this change, the classloader will load the older version of Guava, so MethodNotFound exception might be thrown in some Alluxio  classes due to their needs for the newer version of Guava.

* Add a note for the importance of classpath order.