[#8419] Utilize configuration when populating gRPC options objects (#8421)
* Utilize configuration when populating gRPC options objects

I've also removed the ttl and ttlAction fields from the setAttributes option as it is already included in the common options (#8415). Proto files have been regenerated.

Methods which need options passed on the client side should now utilize the GrpcDefaultOptions class to generate the proper options based off of a configuration object. This will respect users' configuration on the client side. Otherwise options created with getDefaultInstance() don't respect a user's configuration with write type, ttl, etc. This patch will fix that issue.

Fixes #8415
Fixes #8419

* Add license

* fix failing openException test

* Use GrpcDefaultOptions in BaseFileSystem.openFile test

* Add getConf mock for FileSystemContext in FileOutStreamTest

* Rename GrpcDefaultOptions to FileSystemOptions

- Options classes renamed
- methods on FileSystemOptions changed to match FileSystemMasterOptions
- FileSystemMasterOptions exists only for master-only options objects.
- Context classes changed to have defaults(), create(), and mergeFrom() methods to create new instances
- BaseFileSystemTest changed to make sure proper arguments are used to verify on the mocks

* Move FileSystemOptionsTest to proper package, add license

* Move uses of FileSystemOptions back to getDefaultInstance

This should still be ok for tests now that we merge in the proper values on the client side.

Also cleaned up a bit of code with some minor style fixes which aren't caught by checkstyle.

* Fix regression due to IntegrationTestUtils

Because this uses the master client, using the default instance can trigger a sync which can delete files before they are persisted to the UFS. Using FileSystemOptions to utilize the configuration defaults alleviates the issue by setting the sync interval

* Fix for persistCommandTest

Needed to set the write type on the directory so it doesn't get marked as persisted.

* More integration test fixes

Tests which use the underlying client as opposed to the client interface must merge the options from the current configuration or else there will be unexpected behavior.

UfsSyncIntegrationTest needed to set options as opposed to using configuration because once a client is created configuration cannot be changed

* style fixes

* Fix LoadMetadataIntegrationTests

* remove unecessary configuration set on loadMetadata

* add TODO for refactoring mergeFrom in filesystem operation context

* Use getDefaultInstance for clientOps with no options argument

Using getDefaultInstance will save some configuraiton lookups since we now merge within the client method implementation

also fixed a small typo

* remove FileSystemOptions from PathsRestServiceHandler

Because it doesn't specify any specific options other than what's in the configuraiton using FileSystemOptions here is redundant

* Address comments

* restore BufferedBlockInStreamIntegrationTest

The options can be created the same way they were previously

* Fix MigrateDefinition test failures

* style fix