Add support for non-deprecated aliases
An additional Set is now used to mark aliases as non-stable. I chose non-stable as opposed to deprecated because it allows us more flexibility if we want to experiment with other aliases in the future while still providing a warning to users. If the need arises to mark an alias as deprecated we can first add it to the unstable list, and then eventually remove it.

The warning message is adapted from Ubuntu's apt package manager warning (https://askubuntu.com/questions/990823/apt-gives-unstable-cli-interface-warning).

Two minor tests have also been added to verify the new functionality.