ttlAction needs to check for null in DefaultFIleSystemMaster
The setAttribute function should only change the TtlAction if it's specifically set within the options. Otherwise we should avoid the set. Setting the TtlAction with a null param results in the Ttl Action being set to DELETE. If the default is FREE, then a user may be inadvertently setting the action to DELETE by calling SetAttribute.

A test is added which fails before this fix.