Check for null TtlAction in DefaultFileSystemMaster (#8511)
* SetAttribute options are merged when they shouldn't be

In SetAttribute on the filesystem client we previously merged the options
supplied to the client into the FileSystemOptions defaults based on
configuration. The issue is that SetAttribute is a unique operation in
that any value which is sent over the wire we assume was set explicitly by the
user. That's where the issue arose. We were merging default options
into the SetAttributePOptions when we shouldn't have been. The master was
taking the default options and setting the attributes which came from the
defaults.

This change now only explicitly sends a user's options if they are set.
No other options are merged in the setAttribute operation.

In this investigation I also found a bug in setting the TTL and
TtlAction on a file. Both TTL and TtlAction needed to be set in order
for the DefaultFileSystemMaster to actually update the inode. With these
changes this is no longer the case. A user can set Ttl or TtlAction
independently and the master will update the file accordingly.

Lastly, an integration test was added which exposed the original failure
case for both of these issues.

* Fix tests for mock to expect correct options

* ttlAction needs to check for null in DefaultFIleSystemMaster

The setAttribute function should only change the TtlAction if it's specifically set within the options. Otherwise we should avoid the set. Setting the TtlAction with a null param results in the Ttl Action being set to DELETE. If the default is FREE, then a user may be inadvertently setting the action to DELETE by calling SetAttribute.

A test is added which fails before this fix.

* Remove duplicate copy of testMultiSetAttribute

* Update mod time only if ttl or ttlAction was changed

* Add assertions for mod time in testMultiSetAttribute