[#8481] Remove deprecated alluxio filesystem commands (#8482)
* Remove deprecated alluxio fs commands

- Alias commands removed: lsr, rmr, umount
- FileInfoCommand removed
- LoadMetadataCommand removed

* Remove RmrIntegrationTest. Move tests into RmIntegrationTest

* Update LsCommandIntegration test to use ls -R

* Add support for non-deprecated aliases

An additional Set is now used to mark aliases as non-stable. I chose non-stable as opposed to deprecated because it allows us more flexibility if we want to experiment with other aliases in the future while still providing a warning to users. If the need arises to mark an alias as deprecated we can first add it to the unstable list, and then eventually remove it.

The warning message is adapted from Ubuntu's apt package manager warning (https://askubuntu.com/questions/990823/apt-gives-unstable-cli-interface-warning).

Two minor tests have also been added to verify the new functionality.

* Support null unstableAlias and fix various shell constructors

There were quite a few other shells which required the new argument for unStable aliases. Most of them passed null for the command alias maps so I also updated the conditional to check if the unstableAlias Set is null before checking for an item.