Add master hostname not configured exception message
The same or similar message is used throughout the code base
when the proper hostname isn't configured and doesn't include
embedded journal configuration. This change adds a new
message to ExceptionMessage and some utility methods to
generate it due to the number of parameters.

ExceptionMessage exists in core/base which meant I couldn't
use PropertyKeys directly in the message itself. This is why
there are utility methods inside of ConfigurationUtils.