Refactor FileWriteLocationPolicy into BlockLocationPolicy
This refactor involves moving all of the currently implemented policies
into implementations of BlockLocationPolicy. The constructors of each
policy were updated to take a CreateOptions object which holds all of
the necessary information needed to instantiate any of the policies.
From now on, any new implementations should use the
CreateOptions.

The rest of the changes involve trying to remove any reference to
FileWriteLocationPolicy. I've updated the propertyKeys and english
portions of the docs accordingly.

I've also renamed the client options in file_system_master.proto
to be blockRead/Write location policy instead.

One question I came across during this is: Why does copyFromLocal
have it's own configuration property for block location policy which
is separate from the default? It seems out of place.

Places of interest that took extra thought to implement:
- CreateOptions[Test].java
- RoundRobinPolicy[Test].java