Use getFormatsByArgumentIndex instead of getFormats for getMessage
Using getFormatsByIndex() is more preferred because in this
form of message formatting You could have a format "{0} {0}"
and supply the argument "Alluxio" and it would return
"Alluxio Alluxio". Due to the nature of being able to index
arguments, the total number of formats is not necessarily
equivalent to the number of arguments. See the javadoc
for more context

https://docs.oracle.com/javase/7/docs/api/java/text/MessageFormat.html#getFormatsByArgumentIndex()