remove read policy and shards number from open options
By removing these we simply the options object for open RPCs.
We also remove the blockId and blockSize fields on
GetWorkerOptions in favor of a BlockInfo object instead. The
main issue I see arising after this change is that it is not guaranteed that the BlockInfo object is fully populated once
passed into the BlockLocationPolicy#getWorker .