Remove ability to specify write policy in CreateFilePOptions
Removing this option is consistent with the design in the rest of
this refactor. A few tests needed to be modified in order to set
the configuration to use the proper policy instead of simply on
the RPC. Proto files have been regenerated.

 The one main issue is that the copyFromLocal command
doesn't have a simple way of setting the write location policy
separate from what is defined in
alluxio.user.block.write.location.policy. This may degrade
performance in some situations if the worker doesn't have enough space for all blocks in a file.

More thought should probably be taken in order to determine how we may want to mitigate this issue. Right now the
PropertyKey for the copyFromLocal write policy is unused.