Refactor FileWriteLocationPolicy into BlockLocationPolicy (#8552)
* Refactor FileWriteLocationPolicy into BlockLocationPolicy

This refactor involves moving all of the currently implemented policies
into implementations of BlockLocationPolicy. The constructors of each
policy were updated to take a CreateOptions object which holds all of
the necessary information needed to instantiate any of the policies.
From now on, any new implementations should use the
CreateOptions.

The rest of the changes involve trying to remove any reference to
FileWriteLocationPolicy. I've updated the propertyKeys and english
portions of the docs accordingly.

I've also renamed the client options in file_system_master.proto
to be blockRead/Write location policy instead.

One question I came across during this is: Why does copyFromLocal
have it's own configuration property for block location policy which
is separate from the default? It seems out of place.

Places of interest that took extra thought to implement:
- CreateOptions[Test].java
- RoundRobinPolicy[Test].java

* Fix style issues

* remove read policy and shards number from open options

By removing these we simply the options object for open RPCs.
We also remove the blockId and blockSize fields on
GetWorkerOptions in favor of a BlockInfo object instead. The
main issue I see arising after this change is that it is not guaranteed that the BlockInfo object is fully populated once
passed into the BlockLocationPolicy#getWorker .

* fix style

* Remove ability to specify write policy in CreateFilePOptions

Removing this option is consistent with the design in the rest of
this refactor. A few tests needed to be modified in order to set
the configuration to use the proper policy instead of simply on
the RPC. Proto files have been regenerated.

 The one main issue is that the copyFromLocal command
doesn't have a simple way of setting the write location policy
separate from what is defined in
alluxio.user.block.write.location.policy. This may degrade
performance in some situations if the worker doesn't have enough space for all blocks in a file.

More thought should probably be taken in order to determine how we may want to mitigate this issue. Right now the
PropertyKey for the copyFromLocal write policy is unused.

* make copyFromLocal respect its custom write location policy

* Don't close the fsContext in CopyFromLocalCommand

* Remove final traces of FileWriteLocationPolicy

* Remove alluxio.client.block.policy.options.CreateOptions