Track deprecated property keys and warn users of their use
This change aims to expand on the current support we have for
annotated PropertyKeys. Currently property keys are only
annotated as `@java.lang.Deprecated` and any extra info is
included within the generated javadoc. This however isn't
sufficient if we want to warn users at runtime about a key being
deprecated.

This change adds a new custom annotation which allows
property keys to be marked as `@DeprecatedKey` which gives
the ability to set a deprecation message and a deprecated state.

The message is provided to users via log warning if a key is
marked as deprecated. The state of a key will control the severity of the warning. A key with state DEPRECATED will only log a warning is generally a silent unless the user examines the logs. REMOVED will trigger an exception to be thrown and will stop the program from running if a REMOVED key is specified in the configuration.