Fix journal folder deletion error
When Alluxio is used in container environment, it is common that user
mounts an external directory as the journal location in the container.
This will result in error when Alluxio formats embedded journal because
the directory cannot be deleted. This fix solves the issue by not
removing the journal directory but instead removing all the files
inside.

pr-link: Alluxio/alluxio#8638
change-id: cid-91d2b3ea6e0bcf17d9f4bc746609a9adf3f5a2cf