Monitor processes by default in alluxio-start.sh
This change will modify the default behavior of the
alluxio-start.sh script. This change moving forward will make it
easier and more straightforward to start an alluxio cluster and
know when it is ready to start serving requests.

The -m flag has been replaced in alluxio-start.sh with the -a flag.
-a is the logical opposite of what -m performed. Without
specifying any flags, the script will monitor each process to make
sure it has started properly. Otheriwse, if -a (async) flag is
specified, then the processes will start without monitoring and
the script will complete successfully as long as each process
was at least attempted to be started (This is the current
behavior)

The job master/worker functions were also to the alluxio-monitor
script for the (all) and (local) cases.