Support using rocks backups in inode store
This PR leverages rocksdb backup functionality to take faster inode
store checkpoints.

Previously, we would write a checkpoint by iterating over every inode
and serializing it as a journal entry. This is very expensive when most
inodes are stored on-disk in rocksdb. Creating a checkpoint for a system
with 1 billion inodes would take 4+ hours. Restoring from checkpoint
required parsing 1 billion inode journal entries, and writing inodes one
at a time to rocksdb. This would also take 4+ hours.

Instead of traversing the inode tree, we use the rocksdb backup engine
to create a backup, then we turn the backup into a tarball, and use it
for our checkpoint. Restoring is a matter of untarring, then using the
rocksdb backup engine to restore from the backup. I've tested this
approach with 1 billion inodes using a local-disk journal, and it takes
just under 2 hours to create a checkpoint. Restoring from checkpoint
takes just under 1 hour. The size of the checkpoint is 58GB.

This PR makes a few other related changes
1. Change the backup/restore API so that `backup()` takes an
`OutputStream` but `restore()` takes a `CheckpointInputStream`. We do it
this way because all checkpoints should contain the checkpointType ID
parsed by `CheckpointInputStream`. But it's up to individual `backup()`
implementations which checkpointType ID they want to write to their
`OutputStream`.
2. Move checkpoint-related code from `journal` to `journal.checkpoint`
3. Move common functionality shared by `RocksInodeStore` and
`RocksBlockStore` to a new `RocksStore` class. The class is designed so
that it could be used by any other types of rocks storage we might use
in the future.

This PR also addresses some issues that came up during testing:
1. Explicitly close rocks objects when we are finished with them.
Previously we were leaking a lot of rocks iterators and write batches.
Closing these objects frees the off-heap memory that they have
allocated. Garbage collection would also free this memory when the
objects are finalized, but there is no guarantee that this collection
will happen in a timely manner, especially because the garbage collector
isn't aware of the off-heap memory allocation, and may underestimate the
importance of collecting the rocks objects.
2. JournalStateMachine was using the wrong formatting placeholder in
many message (`{}` instead of `%s`)
3. NoopJournaled wasn't writing its checkpoint type. This causes
problems when we try to read 2.0.0-preview checkpoints. We will need to
figure out how we want to support upgrade from 2.0.0-preview to 2.0.0

The new checkpoint format demands updates in `JournalTool` to be able to
read the new checkpoints. This PR is already quite large, so I will
submit that in a followup PR.

pr-link: Alluxio/alluxio#8609
change-id: cid-75bd86bc7a0673b65ad7421edb80449187b8b4e9