Refactor filesystem, fsContext into JobMaster/Worker contexts
This change moves the creation of FileSystem/Context out of
the job definitions and into the respective master/worker
processes. The processes should aim to create clients as few
times as possible because of all the underlying threadpools
and resources that are needed to instantiate a client.

The JobMaster and JobWorkers now pass in a single instance
of a filesystem client+context when selecting executors and
running tasks. This greatly reduces the amount of resources
that the job master and workers consume when there are lots
of jobs being executed.

pr-link: Alluxio/alluxio#8681
change-id: cid-ee3c896e120c6d42087bdeef574e6a8ae05a1580