Refactor JobMaster/WorkerContext into SelectExecutor/RunTaskContext
This refactor introduces the SelectExecutor and
RunTaskContext. The fields of which will be the sole provider
of parameters for job definitions moving forward.

The reason for this changes is that is allows us to keep the
public API for job definitions consistent, but will allow us to
also add new parameters or values to the contexts as desired
in the future. It also allows us to push the creation of the
underlying JobServerContext closer to the beginning of the
callstack within each respective process, so that uneccesary
references aren't created and passed through multiple
methods. Overall it makes JobDefinitions more extensible
and the entire code path cleaner.

The current changes introduce an inheritance model from
JobServerContext which gives both selectExecutors and
runTask access to a filesystem context, client, and a
UfsManager. By inheriting from JobServerContext the
SelectExecutorContext and RunTaskContext can have their
fields changed easily without worrying too much about the
other critical underlying fields that make up the
JobServerContext.