Add "pathConf show" command
Adds "pathConf show" to show path level configurations.

It has two modes:
1. only show configurations for the specific path;
2. show all configurations applicable to the path.

For example, suppose there are two path level configurations:
/a: property1=value1
/a/b: property2=value2
Then in mode 1, only property2=value2 will be shown for /a/b,
but in mode 2, since /a is the prefix of /a/b, both properties will be
shown for /a/b.

pr-link: Alluxio/alluxio#8726
change-id: cid-bfce121ee6de2ea18f1fdf00627d4c4de7889766