Fix authentication with kerberized HDFS
Alluxio master currently initiates login for UFS by calling
connectFromMaster when a UFS is mounted through commandline. However,
this is not called when root ufs is mounted or when nested ufs is
remounted on restart, therefore user has to rely on kinit to enable
Alluxio to talk to Kerberized HDFS.

This change ensures before connecting to a UFS, it always logs in the
user with the credentials from Alluxio properties, which frees up user
from having to use kinit and activates HDFS client relogin mechanism, so
user does not have to periodically kinit to refresh the Kerberos ticket.

pr-link: Alluxio/alluxio#8696
change-id: cid-14a889ec58bfd6e4cab8c35fc2e637ea68bbf252