Improve the start and monitor scripts
alluxio-start was modified so that any time we attempt to start multiple service across nodes, that we use the async flag, and then monitor afterwards.

alluxio-monitor was changed to support clusters in HA mode, and also refactored some of the bash functions. The output was also condensed and colors were modified slightly.