Solve the journal replay issue in concurrent create rename
fixes #8618
There was a time tap in rename() operation between the source path is
visible (createFile(source)can succeed and be journaled) and we journal
the rename entry. This will cause the fail to replay journal issue. In
this PR, we move the journal rename entry step before the making source
path visible step to avoid the above issue.

pr-link: Alluxio/alluxio#8768
change-id: cid-1ef66ce8dea862049e468e183b709836a5256a3d