Implement RocksStore checkpoints using Rocks Checkpoints
Previously we used Rocks BackupEngine to create backups,
and build our checkpoints from those. That was slow because
backups require copying the full metastore. Checkpoints are
faster because they use hard links instead of copying. With
this change, the time needed to take checkpoints and restore
checkpoints is reduced by 40%.