Deflake JournalToolTest
Saw this test fail a new contributor's build.

Previously, the tests could fail if a checkpoint is written
too early, and the journalTool reads that checkpoint. This
PR increases the number of entries required to trigger a
checkpoint, so that we don't take a checkpoint until we
want to.

pr-link: Alluxio/alluxio#8812
change-id: cid-c6eb7e3e7909d85e4a3955a0a9f712036730ece4