Optimize metrics heartbeat thread
For every Alluxio client in a JVM, the client object will instantiate
a new executor with a single thread responsible for heartbeating
metrics to the master. This is quite resource intensive and non-
ideal for this type of task.

This change aims to decrease the resource utilization of the
metrics heartbeat threads within the client. Instead of utilizing
Alluxio's heartbeat threads and a fixed executor service we
instead utilize a static instance of Java's ScheduledExecutorService with a fixed threadpool size of 1.
We schedule a metrics heartbeat task for each client that
gets instantiated with a fixed rate between the execution of
each client's metrics heartbeat. By using this static executor
we save resources by not  needing to instantiate multiple
threads. When a client is created the metrics heartbeat is
scheduled using the single-threaded executor. Instead of
shutting down the executor service when the context is
closed we take the future from
the executor and simple `cancel` it. The scheduled executor
will also hold the thread for the life of the JVM.

The possible downsides to this solution:

- If there are many clients with only a single thread in the
executor, then it is possible that it will be overwhelmed and
metrics heartbeats may arrive at delayed intervals.
- The executor lives for the life of the JVM and never gets
shut down.