Re-design metrics heartbeating
This change aims to completely change how client metrics have been
implemented. Instead of adding the logic into the FileSystemContext
I have created a new class called MetricsHeartbeatContext. This class
solely controls the behavior of metrics heartbeats.

Now the only lines that needs to be called from the
FileSystemContext is the MetricsHeartbeatContext#addHeartbeat
and MetricsHeartbeatContext#removeHeartbeat. The logic for
creating and destroying executors, threadpools, and canceling
heartbeats is handled entirely by this class.

This class supports heartbeating to multiple masters within a single
JVM with multiple client objects. When the first heartbeat is
registered, a new ScheduledExecutorService is created and used
to call heartbeats. No matter the number of applications that are
running within the JVM, there will only ever be one RPC to the
master. That heartbeat will include metrics information for all
clients.

When an application is removed, its Id will be removed from the
heartbeat RPC. When the final heartbeat for a given master is
removed, the RPC will be canceled. If there are no more metrics
RPCs scheduled at that point in time, then the executor service is
shut down until a new metrics heartbeat becomes registered.