Use a single appId per JVM and key with connectDetails
This change addresses two issues:

- Due to the fact that MetricsSystem is still static, we can't
differentiate clients that well with different IDs because all
reported metrics will be the same. To combat this we're only
allowing a single appId to be used which is initialized lazily
based on the first JVM connection.

- Also, instead of using the master address as a key for
different FileSystemContext instances we will use the
ConnectDetails as it is a more accurate representation. I've
also added the #equals and #hashCode methods to the
ConnectDetails interface to ensure that implementations are
compatible as Map keys.