Remove duplication of ClientContext and MasterClientContexts in FileSystemContext.
MasterClientContexts in the FileSystemMasterClientPool and BlockMasterClientPool duplicate references to members in ClientContext, but this causes trouble of maintaining consistency of updated configurations in those contexts.

This commit consolidates all contexts to use a single instance of MasterClientContext, so that updates to configurations in that context can be visible to all threads.