Track metrics over worker and client restarts
This PR improves the Metrics collection on the master and
workers by allowing cluster-aggregated metrics on the master
to reflect accurate values even when workers (or clients)
restart.

In order to support properly tracking metrics over restarts we
need to know which values will differ over restarts. Alluxio's
MetricsSystem uses 4 types of metrics: GAUGE, COUNTER,
METER, and TIMER. COUNTER metrics are the only ones
which change over restarts. The rest vary with time, and can
safely assume the same behavior they had previously.

This change adds MetricType, and updates the client and
master behavior such that when the client or worker sends
metrics over the wire, that the metrics reported for COUNTER
will only include the incremental values that the given metrics
should change, rather than the absolute value since the
process started. This allows the master to track metrics for
the life of the process and its workers, even if workers restart.