Implement sink for journal entries
The journal sink is a new interface for the journal system to write
journal entries to. The journal sink is useful for examining journal
entries applied to the master. Only LOG journal entries are appended to
the sink, so no checkpoint entries are written to the sink.

The journal system writes entries to the journal sink, as the leader and
also as the standby. If the journal is the leader, the journal system
appends to the sink when writing/flushing to the journal. If the journal
is a standby, the journal system appends to the sink when replaying the
journal to catch up.

Journal sink implementations must not block any of the methods, since
they are called within the journal system, and any delay will negatively
impact operational clients on the leader master.

pr-link: Alluxio/alluxio#8995
change-id: cid-da3b0e82eeee4398f9a699eaeef820694c1df458