Remove implicit inode locking
Previously, we would skip taking "unnecessary" locks on inodes and
edges. A lock was considered unnecessary if one of its ancestors was
already write-locked, e.g. if we've write-locked /a/b, we don't need to
acquire any more locks to lock /a/b/c/d. This is what we called implicit
locking.

Implicit locking works well when all locks are taken starting from the
root. However, we sometimes want to lock individual inodes without
needing to lock their ancestors. In particular, CachingInodeStore needs
to evict inodes from memory to RocksDB, which requires locking the
inodes directly (locking their ancestors would be too expensive).

To support random locks, we remove the concept of implicit locking. With
this PR, we replace implicit locks with explicit write locks. For the
above example, we will now take write locks on all inodes and edges
leading from /a/b to /a/b/c/d.

One new invariant we follow is that all lock lists contain some number
of read-locked entries followed by some number of write-locked entries.
In particular, we never take a read lock following a write lock. So if
you've write-locked /a/b and then request to extend the path with a
read-lock on /a/b/c, you will actually take a write lock.

pr-link: Alluxio/alluxio#9042
change-id: cid-44a7be7b7ed499c43f4b6b5ca0b6bcdc5f3a066d