Track deprecated property keys and warn users of their use
This change aims to expand on the current support we have for
annotated PropertyKeys. Currently, property keys are only
annotated as `@java.lang.Deprecated` and any extra info is
included within the generated javadoc. This however isn't
sufficient if we want to warn users at runtime about a key being
deprecated or removed.

This change adds two new custom annotations which allows
property keys to be marked as `@alluxio.conf.Deprecated` which
gives the ability to set a deprecation message. The other one is
`@alluxio.conf.Removed` which is similar to deprecated but
denotes PropertyKeys that have been completely removed from
use.

The messages within the annotations are provided to users via log
warning or an exception message if a key is marked as `@Deprecated`
or `@Removed`. When either annotation is provided on a key the
configuration validation step is responsible for logging a warning in
the case of `@Deprecated` or throwing a RuntimeException for
`@Removed`. Keys that reside in `RemovedKeys.java` are there for
historical purposes and shouldn't be used.

pr-link: Alluxio/alluxio#8616
change-id: cid-af3646499e8e3f589e448d83031160b3ffac05d9