Add Pinning to specific tier
There are several things in this PR.

Change Inode metadata to include a set of strings that are the desired
pinning medium for the file.
Pass a desired pinning storage medium type to loadBlock method of
JobUtils.java.
This storage medium type eventually gets passed to the createBlock
method on the worker, and worker finds a storage medium which matches
the desired storage medium type.
Change the blockInfo in fileInfo to contain medium type information.

TODO (defer to phase 2):
BlockHeartBeatReport related changes

pr-link: Alluxio/alluxio#9084
change-id: cid-4f9ba15be23d9228b8ccf78fac609a529d648c3b